adding delete and edit with authentication

+++ src/actions/actions.ts
@@ -72,13 +72,37 @@ export async function addPet(pet: unknown) {
 }
 
 export async function editPet(petId: unknown, newPetData: unknown) {
+  //authentication
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  //validation
   const validatedPetId = petIdSchema.safeParse(petId);
   const validatedPet = petFormSchema.safeParse(newPetData);
+
   if (!validatedPetId.success || !validatedPet.success) {
     return {
       message: "Invalid pet data",
     };
   }
+
+  //authorization
+  const pet = await prisma.pet.findUnique({
+    where: { id: validatedPetId.data },
+  });
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized to edit this pet",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: { id: validatedPetId.data },
@@ -99,6 +123,28 @@ export async function deletePet(petId: unknown) {
       message: "Invalid pet data",
     };
   }
+
+  //authorization
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+  const pet = await prisma.pet.findUnique({
+    where: { id: validatedPetId.data },
+  });
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "You are not authorized to delete this pet",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: { id: validatedPetId.data },
